use singleton ObjectMapper
Reduces some memory footprint as single reused object mapper is used
instead of creating new for each output stream, key, header and value
serializer.
