+++ src/main/java/io/aiven/kafka/connect/common/output/jsonwriter/JsonLinesOutputStreamWriter.java
@@ -20,54 +20,17 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
-import java.util.Set;
 
-import org.apache.kafka.connect.sink.SinkRecord;
-
-import io.aiven.kafka.connect.common.output.OutputStreamWriter;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.JsonNodeFactory;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-
-public class JsonLinesOutputStreamWriter implements OutputStreamWriter {
+public class JsonLinesOutputStreamWriter extends AbstractJsonOutputStreamWriter {
     private static final byte[] RECORD_SEPARATOR = "\n".getBytes(StandardCharsets.UTF_8);
-
-    private final Map<String, OutputFieldBuilder> fieldBuilders;
-    private final boolean envelopeEnabled;
-    private final ObjectMapper objectMapper;
 
     JsonLinesOutputStreamWriter(final Map<String, OutputFieldBuilder> fieldBuilders,
                                        final boolean envelopeEnabled) {
-        this.fieldBuilders = fieldBuilders;
-        this.envelopeEnabled = envelopeEnabled;
-        this.objectMapper = new ObjectMapper();
-        objectMapper.setNodeFactory(JsonNodeFactory.withExactBigDecimals(true));
+        super(fieldBuilders, envelopeEnabled);
     }
 
     @Override
     public void writeRecordsSeparator(final OutputStream outputStream) throws IOException {
         outputStream.write(RECORD_SEPARATOR);
-    }
-
-    @Override
-    public void writeOneRecord(final OutputStream outputStream, final SinkRecord record) throws IOException {
-        outputStream.write(objectMapper.writeValueAsBytes(getFields(record)));
-    }
-
-    private JsonNode getFields(final SinkRecord record) throws IOException {
-        if (envelopeEnabled) {
-            final ObjectNode root = JsonNodeFactory.instance.objectNode();
-            final Set<Map.Entry<String, OutputFieldBuilder>> entries = fieldBuilders.entrySet();
-            for (final Map.Entry<String, OutputFieldBuilder> entry : entries) {
-                final JsonNode node = entry.getValue().build(record);
-                root.set(entry.getKey(), node);
-            }
-            return root;
-        } else {
-            // envelope can be disabled only in case of single field
-            return fieldBuilders.entrySet().iterator().next().getValue().build(record);
-        }
     }
 }
